move emailcontroller tests to editProfileController
